feat: 新增在安装之后，自动验证是否是已安装Chromium，如果没有，则自动安装
