+++ package.json
@@ -8,7 +8,8 @@
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "generate": "all-contributors generate"
+    "generate": "all-contributors generate",
+    "postinstall": "./scripts/chromium-checker"
   },
   "keywords": [
     "sms",
